+++ docs/source/changelog.rst
@@ -43,7 +43,9 @@ Changelog
 * Changed the return type of :meth:`Model.save <pyairtable.orm.Model.save>`
   from ``bool`` to :class:`~pyairtable.orm.SaveResult`.
   - `PR #387 <https://github.com/gtalarico/pyairtable/pull/387>`_
-* Added support for `Upload attachment <https://airtable.com/developers/web/api/upload-attachment>`_.
+* Added support for `Upload attachment <https://airtable.com/developers/web/api/upload-attachment>`_
+  via :meth:`Table.upload_attachment <pyairtable.Table.upload_attachment>`
+  or :meth:`AttachmentsList.upload <pyairtable.orm.lists.AttachmentsList.upload>`.
 
 2.3.3 (2024-03-22)
 ------------------------